0x01/round_2/2.web/resources/webapp-2/requirements.txt
Flask==2.2.2
pyjwt==2.6.0
mysql-connector-python==8.0.31
mysql-connector==2.2.9